requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.3.0=pypi_0
aiohttp=3.8.3=py39h02fc5c5_1
aiosignal=1.3.1=pyhd8ed1ab_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.2.0=pyh71513ae_0
blinker=1.5=pyhd8ed1ab_0
brotlipy=0.7.0=py39h02fc5c5_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.12.7=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=py39h7e6b969_3
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
contourpy=1.0.6=pypi_0
cryptography=39.0.0=py39haa0b8cc_0
cvzone=1.5.6=pypi_0
cycler=0.11.0=pypi_0
dropbox=11.36.0=pypi_0
flask=2.2.2=pypi_0
flask-wtf=1.0.1=pypi_0
flatbuffers=2.0.7=pypi_0
fonttools=4.38.0=pypi_0
frozenlist=1.3.3=py39h02fc5c5_0
gast=0.4.0=pyh9f0ad1d_0
google-auth=2.15.0=pypi_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.51.1=pypi_0
h5py=3.7.0=pypi_0
hdf5=1.10.6=nompi_h0fc092c_1114
idna=3.4=pyhd8ed1ab_0
imbalanced-learn=0.10.1=pypi_0
importlib-metadata=6.0.0=pyha770c72_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pyhd8ed1ab_0
keras=2.7.0=pypi_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h127bd45_0
libabseil=20220623.0=cxx17_h28b99d4_6
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libclang=14.0.6=pypi_0
libcurl=7.87.0=hbe9bab4_0
libcxx=14.0.6=h2692d47_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=11_3_0_hd922786_27
libgfortran5=11.3.0=hdaf2cc0_27
libgrpc=1.51.1=h503f348_0
liblapack=3.9.0=16_osxarm64_openblas
libllvm11=11.1.0=hfa12f05_5
libnghttp2=1.51.0=hd184df1_0
libopenblas=0.3.21=openmp_hc731615_3
libprotobuf=3.21.12=hb5ab8b9_0
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=hb80f160_3
libzlib=1.2.13=h03a7124_4
llvm-openmp=15.0.6=h7cfbb63_0
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py39h02fc5c5_2
matplotlib=3.6.2=pypi_0
multidict=6.0.4=py39h02fc5c5_0
ncurses=6.3=h07bb92c_1
numpy=1.24.1=pypi_0
oauthlib=3.2.2=pyhd8ed1ab_0
openai=0.27.7=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1s=h03a7124_1
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=22.0=pypi_0
pandas=1.5.2=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=pyhd8ed1ab_0
ply=3.11=pypi_0
protobuf=3.20.3=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
pytelegrambotapi=4.12.0=pypi_0
python=3.9.15=h2d96c93_0_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.9=3_cp39
pytz=2022.7=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
re2=2022.06.01=h9a09cb3_1
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scikit-learn=1.0.2=pypi_0
scipy=1.9.1=py39h737da60_0
setuptools=65.6.3=pypi_0
six=1.16.0=pypi_0
stone=3.3.1=pypi_0
tensorboard=2.11.0=pypi_0
tensorboard-data-server=0.6.1=py39haa0b8cc_4
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow-deps=2.7.0=0
tensorflow-estimator=2.7.0=pypi_0
tensorflow-macos=2.7.0=pypi_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.1.0=pyh8a188c0_0
tk=8.6.12=he1e0b03_0
tqdm=4.65.0=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022g=h191b570_0
urllib3=1.26.13=pyhd8ed1ab_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.38.4=pypi_0
wrapt=1.14.1=pypi_0
wtforms=3.0.1=pypi_0
xz=5.2.6=h57fd34a_0
yarl=1.8.2=py39h02fc5c5_0
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.13=h03a7124_4
